move stream ids to separate header
